requirements.txt
fastapi
uvicorn[standard]
pandas-market-calendars
pandas
google-cloud